Gestion de Usuario Completo, Cambio de Contraseña, Landing Home

+++ src/app/componentes/table/table-products/table-products.component.ts
@@ -51,14 +51,16 @@ export class TableProductsComponent {
 
   //variables para obtener y cargar productos
   products: Product[] = [];
-  inventarios:Inventory[] = [];
+  inventarios: Inventory[] = [];
   productosCompletos: ProductWithInventory[] = [];
+  productoEditable: any = {};
+  editarProductoModalVisible = false;
 
   //variables para registrar un producto
   brands: Brand[] = [];
   categories: Category[] = [];
   warranties: Warranty[] = [];
-  
+
   nuevoProducto: CreateProduct = {
     brand_id: 0,
     category_id: 0,
@@ -76,6 +78,31 @@ export class TableProductsComponent {
   stock: number = 0;
   price_usd: number = 0;
 
+  abrirModalEditarProducto(productId: number): void {
+    const productosCompleto = this.productosCompletos.find(p => p.id === productId);
+    if (!productosCompleto) {
+      console.error('Producto no encontrado para editar');
+      return;
+    }
+
+    this.productoEditable = {
+      id: productosCompleto.id,
+      name: productosCompleto.name,
+      description: productosCompleto.description,
+      technical_specifications: productosCompleto.technical_specifications,
+      brand_id: productosCompleto.brand_id,
+      category_id: productosCompleto.category_id,
+      warranty_id: productosCompleto.warranty_id,
+      model_3d_url: productosCompleto.model_3d_url,
+      ar_url: productosCompleto.ar_url,
+      image: null, // file solo si se cambia
+      inventory_id: productosCompleto.inventory_id,
+      price_usd: productosCompleto.price_usd,
+      stock: productosCompleto.stock
+    };
+    this.editarProductoModalVisible = true;
+  }
+
 
   constructor(public productos: ProductosService,
     private authService: AuthService,
@@ -84,8 +111,8 @@ export class TableProductsComponent {
 
   ngOnInit() {
     this.cargarDatosCompletos();
-    this.cargarFormOpciones();   
-    console.log("productos cargados:",this.productos);
+    this.cargarFormOpciones();
+    console.log("productos cargados:", this.productos);
   }
 
   cargarProductos() {
@@ -104,44 +131,51 @@ export class TableProductsComponent {
       },
       error: (err) => console.error('Error al obtener inventario', err),
     });
-  }    
+  }
 
   cargarDatosCompletos() {
     this.productos.obtenerProductos().subscribe({
       next: (resProd) => {
         const productos = resProd.items;
-
+  
         this.productos.obtenerInventarioCompleto().subscribe({
           next: (resInv) => {
             const inventarios = resInv.items;
-            console.log("productos: ",productos);
-            // Unimos productos con inventario por ID
+  
+            console.log("Productos recibidos:", productos);
+  
+            // Unimos productos con inventario por product_id
             this.productosCompletos = productos.map(producto => {
               const inv = inventarios.find(i => i.product_id === producto.id);
-              
+              if (!inv) return null;
+            
               return {
                 id: producto.id,
                 name: producto.name,
                 active: producto.active,
                 image_url: producto.image_url,
                 category: producto.category.name || '',
+                category_id: producto.category.id, 
+                brand_id: producto.brand.id,
+                warranty_id: producto.warranty.id,
                 technical_specifications: producto.technical_specifications || '',
                 description: producto.description || '',
-                price_usd: inv?.price_usd ?? 0,
-                stock: inv?.stock ?? 0,
+                price_usd: inv.price_usd,
+                stock: inv.stock,
+                inventory_id: inv.id,
+                model_3d_url: producto.model_3d_url || '',
+                ar_url: producto.ar_url || ''
               };
-              
-              
-            });
-            
+            }).filter(p => p !== null);; 
             
           },
-          error: (err) => console.error("Error al obtener inventario", err),
+          error: (err) => console.error("❌ Error al obtener inventario", err),
         });
       },
-      error: (err) => console.error("Error al obtener productos", err),
+      error: (err) => console.error("❌ Error al obtener productos", err),
     });
   }
+  
 
   cargarFormOpciones(): void {
     this.productos.getBrands().subscribe({
@@ -210,45 +244,92 @@ export class TableProductsComponent {
           stock: this.stock,
           price_usd: this.price_usd
         };
-  
+
         this.productos.createInventory(inventario).subscribe({
           next: () => {
-            this.noti.success('Registro exitoso','Producto registrado exitosamente');
+            this.noti.success('Registro exitoso', 'Producto registrado exitosamente');
             this.nuevoProductoModalVisible = false;
             // Recargar tabla, etc...
             this.cargarDatosCompletos();
           },
           error: (err) => {
             console.error('Error al registrar inventario:', err);
-            this.noti.error('No se pudo regitrar producto con inventario','Hubo un error al registrar el producto');
+            this.noti.error('No se pudo regitrar producto con inventario', 'Hubo un error al registrar el producto');
           }
         });
       },
       error: (err) => {
         console.error('Error al registrar producto', err);
-        this.noti.error('No se pudo regitrar producto','Hubo un error al registrar el producto');
+        this.noti.error('No se pudo regitrar producto', 'Hubo un error al registrar el producto');
       }
     });
-}
-
-
-
-
-  actualizarProducto(updatedProduct: Product): void {
-    this.productos.actualizarProducto(updatedProduct).subscribe({
-      next: () => this.cargarProductos(),
-      error: (err) => console.error('Error al actualizar el producto', err),
-    })
   }
 
-  eliminarProducto(id: number): void {
-    this.productos.eliminarProducto(id).subscribe({
-      next: () => this.cargarProductos(),
-      error: (err) => console.error('Error al eliminar el producto', err),
+
+  onEditFileSelected(event: any): void {
+    const file = event.target.files[0];
+    if (file) this.productoEditable.image = file;
+  }
+
+
+  /*Funcion para abrir el modal con los datos cargados*/
+  editarProducto(): void {
+    const formData = new FormData();
+    formData.append('name', this.productoEditable.name);
+    formData.append('description', this.productoEditable.description);
+    formData.append('technical_specifications', this.productoEditable.technical_specifications);
+    formData.append('brand_id', this.productoEditable.brand_id.toString());
+    formData.append('category_id', this.productoEditable.category_id.toString());
+    formData.append('warranty_id', this.productoEditable.warranty_id.toString());
+    formData.append('model_3d_url', this.productoEditable.model_3d_url || '');
+    formData.append('ar_url', this.productoEditable.ar_url || '');
+    if (this.productoEditable.image) formData.append('image', this.productoEditable.image);
+
+
+    this.productos.editarProducto(this.productoEditable.id, formData).subscribe({
+      next: () => {
+        const invData = {
+          stock: this.productoEditable.stock,
+          price_usd: this.productoEditable.price_usd
+        };
+        this.productos.actualizarInventario(this.productoEditable.inventory_id, invData).subscribe({
+          next: () => {
+            this.noti.success('Producto actualizado', 'Producto e Inventario Actualizados');
+            this.editarProductoModalVisible = false;
+            console.log("Productos con inventario actualizados:", this.productosCompletos);
+            this.cargarDatosCompletos();
+          },
+          error: (err) => {
+            console.error('Error al actualizar inventario:', err);
+            this.noti.error('Error', 'No se pudo actualizar el inventario');
+          }
+        });
+      },
+      error: (err) => {
+        console.error('Error al actualizar producto', err);
+        this.noti.error('Error', 'No se pudo actualizar el producto');
+      }
     });
   }
 
-    // agregarProducto(newProduct: Product): void {
+
+
+  eliminarProducto(id: number): void {
+    if (confirm('¿Estás seguro de que deseas eliminar este producto?')) {
+      this.productos.eliminarProducto(id).subscribe({
+        next: () => {
+          this.noti.success('Producto eliminado', 'El producto fue eliminado correctamente');
+          this.cargarDatosCompletos(); // mejor usar cargarDatosCompletos si usás inventario también
+        },
+        error: (err) => {
+          console.error('Error al eliminar el producto', err);
+          this.noti.error('Error', 'No se pudo eliminar el producto');
+        }
+      });
+    }
+  }
+
+  // agregarProducto(newProduct: Product): void {
   //   this.productos.agregarProductos(newProduct).subscribe({
   //     next: () => this.cargarProductos(),
   //     error: (err) => console.error('Error al agregar el producto', err),
